pyproject.toml
[project]
name = "request-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.14.1",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "transformers>=4.51.3",
    "uvicorn>=0.34.2",
    "xlrd>=2.0.1",
]
